requirements.txt
aiofiles
allure-pytest
allure-python-commons
asgiref
atomicwrites
attrs
blinker
Brotli
certifi
cffi
chardet
charset-normalizer
click
colorama
colorlog
cryptography
DingtalkChatbot
et-xmlfile
execnet
Faker
Flask
h11
h2
hpack
httptools
hyperframe
idna
iniconfig
itchat
itsdangerous
Jinja2
jsonpath
kaitaistruct
ldap3
MarkupSafe
mitmproxy
msgpack
multidict
openpyxl
packaging
passlib
pluggy
protobuf
publicsuffix2
py
pyasn1
pycparser
pydivert
PyMySQL
pyOpenSSL
pyparsing
pyperclip
pypng
PyQRCode
pytest
pytest-forked
pytest-xdist
python-dateutil
pywin32
PyYAML
requests
requests-toolbelt
ruamel.yaml
ruamel.yaml.clib
sanic
sanic-routing
six
sortedcontainers
text-unidecode
toml
tornado
urllib3
urwid
websockets
Werkzeug
wsproto
xlrd
xlutils
xlwings
xlwt
pyDes
crypto
redis
pydantic
